Plexams.Types.Exam

Documentation

data Exam

Constructors

Exam

Fields

anCode :: Ancode

Anmeldecode Prüfungsamt

name :: String

Name der Prüfung

lecturer :: Person

Prüfer

duration :: Duration

Dauer der Prüfung in Minuten

rooms :: [Room]

Liste der Räume in denen die Prüfung statt findet

plannedByMe :: Bool

False bei Prüfungen, die zwar mit erfasst werden, aber nicht geplant werden können

reExam :: Bool

True bei einer Wiederholungsklausur

groups :: [Group]

Studierendengruppen die an der Prüfung teilnehmen

examType :: ExamType

Typ der Prüfung aus ZPA

slot :: Maybe (Int, Int)

(Tag, Slot)

registeredStudents :: [StudentWithRegs]

registeredStudentsCount :: Integer

registeredGroups :: [RegisteredGroup]

conflictingAncodes :: Map Ancode Integer

Ancode maps to number of students with this conflict

handicapStudents :: [StudentWithRegs]

sameRoom :: [Ancode]

sameSlot :: [Ancode]

shareRoom :: Bool

registrations :: Exam -> Integer

isScheduled :: Exam -> Bool

isUnscheduled :: Exam -> Bool

isGOExam :: Exam -> Bool

withHandicaps :: Exam -> Bool

handicapStudentsNeedsRoomAlone :: Exam -> [StudentWithRegs]

withHandicapsNeedsRoomAlone :: Exam -> Bool

notPlannedByMe :: [Ancode] -> Exam -> Exam

seatsMissing :: Exam -> Integer

showMinimal :: Exam -> String