Plexams.Types.Rooms

Documentation

data Room

Constructors

Room

Fields

roomID :: RoomID

Raum-Nr, z.B. "R3.014"

maxSeats :: Integer

maximale Anzahl an Prüfungsplätzen

deltaDuration :: Duration

falls der Raum für NTA genutzt wird, Anzahl der Minuten die die Prüfung länger dauert

invigilator :: Maybe Invigilator

Aufsicht

reserveRoom :: Bool

True, Raum ist eingeplant, wird aber nicht im ZPA veröffentlicht

handicapCompensation :: Bool

True Raum für NTA

studentsInRoom :: [StudentWithRegs]

seatsPlanned :: Room -> Integer

data PlannedRoomWithSlots

Constructors

PlannedRoomWithSlots

Fields

plannedRoomID :: String

plannedRoomDaysAndSlots :: [PlannedRoomSlot]

data PlannedRoomSlot

Constructors

PlannedRoomSlot

Fields

plannedRoomDay :: String

plannedRoomDayNumber :: Int

plannedRoomSlots :: [(Int, String)]